% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{BoPModels}
\alias{BoPModels}
\title{BoPModels Function}
\usage{
BoPModels(mydata = NULL)
}
\arguments{
\item{mydata}{No default.  If null is passed, it will load data from "updatedbankingdata.csv under the data folder}
}
\description{
This function will take the dataframes and build glm, random forest and svm models. The prediction data will be
saved in files called model_prediction_glm.csv," "model_prediction_svm.csv," and "model_prediction_svm.csv"
}
\examples{
BoPModels()
}
\keyword{models}
