% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{loadData}
\alias{loadData}
\title{loadData Function}
\usage{
loadData(infile = "data/bank-additional-full.csv")
}
\arguments{
\item{infile}{Defaults to "bank-additional-full.csv" under the data folder}
}
\value{
a dataframe of training data and  a dataframe of testing data
}
\description{
This function allows you to update the "updatedbankingdata.csv" from the infile of your choice.  The infile should be ; separated csv files. The dataframe will be returned.
}
\examples{
loadData()
}
\keyword{loaddata}
